refactor: Custom Exception, 계층적 구조로 변경
test: GameServiceInitTest, 이름 길이 예외 테스트 실패 해결
