+++ src/main/java/racingcar/exception/WrongArgumentException.java
@@ -1,7 +1,5 @@
 package racingcar.exception;
 
-import racingcar.exception.status.ExceptionStatus;
-
 public class WrongArgumentException extends IllegalArgumentException {
 
     public WrongArgumentException(final ExceptionStatus exceptionStatus) {
