+++ src/main/java/racingcar/exception/car/WrongCarNameException.java
@@ -1,0 +1,12 @@
+package racingcar.exception.car;
+
+import racingcar.exception.ExceptionStatus;
+import racingcar.exception.WrongArgumentException;
+
+public class WrongCarNameException extends WrongArgumentException {
+
+    public WrongCarNameException(ExceptionStatus exceptionStatus) {
+        super(exceptionStatus);
+    }
+
+}
