+++ src/main/java/racingcar/exception/reader/WrongReaderException.java
@@ -1,0 +1,12 @@
+package racingcar.exception.reader;
+
+import racingcar.exception.ExceptionStatus;
+import racingcar.exception.WrongArgumentException;
+
+public class WrongReaderException extends WrongArgumentException {
+
+    public WrongReaderException(ExceptionStatus exceptionStatus) {
+        super(exceptionStatus);
+    }
+
+}
