+++ src/main/java/racingcar/exception/round/WrongRoundCountException.java
@@ -1,0 +1,12 @@
+package racingcar.exception.round;
+
+import racingcar.exception.ExceptionStatus;
+import racingcar.exception.WrongArgumentException;
+
+public class WrongRoundCountException extends WrongArgumentException {
+
+    public WrongRoundCountException(ExceptionStatus exceptionStatus) {
+        super(exceptionStatus);
+    }
+
+}
